refactor(materializer): extract transcript-patch conversation strategies (file #3)
Lift the 6 transcript-patch conversation fast paths (array-patch / streaming /
prefix-patch / append / patch+append / truncate) + their tail helpers into
transcript-patch-conversations.ts; dispatcher imports them. Net 74/74, check 0.
